+++ example_supabase/lib/brick/adapters/pizza_adapter.g.dart
@@ -49,7 +49,7 @@ class PizzaAdapter extends OfflineFirstWithSupabaseAdapter<Pizza> {
   @override
   final defaultToNull = true;
   @override
-  final Map<String, RuntimeSqliteColumnDefinition> fieldsToSqliteColumns = {
+  final fieldsToSupabaseColumns = {
     'id': const RuntimeSupabaseColumnDefinition(
       association: false,
       associationForeignKey: 'null',