+++ packages/brick_offline_first_with_supabase/example/lib/brick/adapters/mounty_adapter.g.dart
@@ -58,7 +58,7 @@ class MountyAdapter extends OfflineFirstWithSupabaseAdapter<Mounty> {
   @override
   final defaultToNull = true;
   @override
-  final Map<String, RuntimeSqliteColumnDefinition> fieldsToSqliteColumns = {
+  final fieldsToSupabaseColumns = {
     'name': const RuntimeSupabaseColumnDefinition(
       association: false,
       associationForeignKey: 'null',