+++ packages/brick_offline_first_with_supabase_build/test/offline_first_generator/test_table_name_undefined.dart
@@ -41,11 +41,11 @@ class UndefinedCamelizedNameAdapter
   UndefinedCamelizedNameAdapter();
 
   @override
-  final tableName = 'undefined_camelized_names';
+  final supabaseTableName = 'undefined_camelized_names';
   @override
   final defaultToNull = true;
   @override
-  final Map<String, RuntimeSqliteColumnDefinition> fieldsToSqliteColumns = {
+  final fieldsToSupabaseColumns = {
     'someLongField': const RuntimeSupabaseColumnDefinition(
       association: false,
       associationForeignKey: 'null',