+++ packages/brick_supabase/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+## 0.1.1
+
+- Resolve `QuerySupabaseTransformer` bugs
+- Fix table name collision in offline adapter (#416)
+
 ## 0.1.0
 
 - Alpha release